Move tests from pages folder
