+++ src/__tests__/pages/index.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@test/test-helper'
-import Home, { getStaticProps } from '../index'
+import Home, { getStaticProps } from '../../pages/index'
 
 jest.mock('@sections/home/index', () => () => <div>Home</div>)
 jest.mock('@functions/fetchContentful', () => ({ type }) => type)
